Remove validation on encode
